requirements.txt
Flask_SQLAlchemy>=2.4,<=3.0
pytest
pyyaml
flask
marshmallow